Move UserGender to within User. Fix name consistency
